// src/app/interfaces/activities.ts
export interface ActivityUnity {
    position: number;
    activity: string;
    type: string;
    unity: string;
  }

export interface ActivityInfo {
  id: number;
  activity_name: string;
  shortcut: string;
}

export interface FieldType {
  typeName: string;
}

/*export interface UserActivities {
  user: string;
  activityName: string;
  activated: boolean;
  labelName?: string;
}*/